refactor: separate android & ios bridge
