+++ js-miniapp-bridge/package.json
@@ -28,9 +28,9 @@
   "scripts": {
     "lint": "gts check",
     "clean": "gts clean",
-    "compile": "tsc -p .",
+    "compile": "tsc -p . && browserify ./build/src/android-bridge.js > build/bridge.js",
     "prebuild": "npm run clean && npm run lint",
-    "build": "npm run lint && npm run compile && npm run test",
+    "buildBridge": "npm run lint && npm run compile && npm run test",
     "fix": "gts fix",
     "test": "mocha test/**/*.spec.ts",
     "coverage": "nyc npm test"
