+++ js-miniapp-bridge/test/setup.js
@@ -1,0 +1,6 @@
+'use strict';
+
+const chai = require('chai');
+
+chai.use(require('chai-as-promised'))
+chai.use(require('sinon-chai'));
